# mentorme_backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Mentor Me API
  description: >
    Backend API for the Mentor Me tutor–student marketplace MVP.
  version: 0.1.0
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:4000
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [STUDENT, TUTOR, ADMIN]
        status:
          type: string
          enum: [ACTIVE, PENDING, SUSPENDED]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    StudentProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
        gradeLevel:
          type: string
        goals:
          type: string
          nullable: true
        preferredSubjects:
          type: array
          items:
            type: string
        notes:
          type: string
          nullable: true
    TutorProfile:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        bio:
          type: string
          nullable: true
        education:
          type: string
          nullable: true
        certificates:
          type: array
          items:
            type: string
        yearsOfExperience:
          type: integer
        hourlyRateMin:
          type: number
          nullable: true
        hourlyRateMax:
          type: number
          nullable: true
        teachingModes:
          type: array
          items:
            type: string
        city:
          type: string
          nullable: true
        district:
          type: string
          nullable: true
        verified:
          type: boolean
        trustScore:
          type: number
        averageRating:
          type: number
        totalCompletedBookings:
          type: integer
        totalReviews:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Booking:
      type: object
      properties:
        id:
          type: string
        classId:
          type: string
        studentId:
          type: string
        tutorId:
          type: string
        status:
          type: string
          enum: [PENDING, CONFIRMED, TRIAL, CANCELLED, COMPLETED]
        isTrial:
          type: boolean
        requestedHoursPerWeek:
          type: integer
        startDateExpected:
          type: string
          format: date
        noteFromStudent:
          type: string
          nullable: true
        cancelReason:
          type: string
          nullable: true
        cancelledBy:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fullName, email, phone, password, role]
              properties:
                fullName:
                  type: string
                email:
                  type: string
                phone:
                  type: string
                password:
                  type: string
                role:
                  type: string
                  enum: [STUDENT, TUTOR]
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  accessToken:
                    type: string
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login with email or phone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [emailOrPhone, password]
              properties:
                emailOrPhone:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  accessToken:
                    type: string
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  studentProfile:
                    $ref: "#/components/schemas/StudentProfile"
                  tutorProfile:
                    $ref: "#/components/schemas/TutorProfile"
...
  /api/classes:
    post:
      tags: [Class]
      summary: Create class listing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - subjectId
                - title
                - description
                - pricePerHour
                - locationType
              properties:
                subjectId:
                  type: string
                title:
                  type: string
                description:
                  type: string
                targetGrade:
                  type: string
                pricePerHour:
                  type: number
                locationType:
                  type: string
                  enum: [ONLINE, AT_STUDENT, AT_TUTOR]
                city:
                  type: string
                district:
                  type: string
      responses:
        "201":
          description: Created class
...
  /api/bookings:
    post:
      tags: [Booking]
      summary: Create booking request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - classId
                - requestedHoursPerWeek
                - startDateExpected
              properties:
                classId:
                  type: string
                isTrial:
                  type: boolean
                requestedHoursPerWeek:
                  type: integer
                startDateExpected:
                  type: string
                  format: date
                noteFromStudent:
                  type: string
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    get:
      tags: [Booking]
      summary: List bookings (role-aware)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: List of bookings
...
  /api/admin/tutors/pending:
    get:
      tags: [Admin]
      summary: List tutors pending verification
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Pending tutors
  /api/admin/tutors/{id}/verify:
    patch:
      tags: [Admin]
      summary: Verify or reject tutor
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [approved]
              properties:
                approved:
                  type: boolean
                note:
                  type: string
      responses:
        "200":
          description: Result message
  /api/admin/tutors/{id}/ban:
    patch:
      tags: [Admin]
      summary: Ban or unban tutor
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [banned]
              properties:
                banned:
                  type: boolean
      responses:
        "200":
          description: Result message
  /api/admin/students/{id}/ban:
    patch:
      tags: [Admin]
      summary: Ban or unban student
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [banned]
              properties:
                banned:
                  type: boolean
      responses:
        "200":
          description: Result message
  /api/admin/bookings:
    get:
      tags: [Admin]
      summary: Admin booking list
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: tutorId
          schema:
            type: string
        - in: query
          name: studentId
          schema:
            type: string
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Filtered bookings
  /api/admin/classes:
    get:
      tags: [Admin]
      summary: Admin class list
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: tutorId
          schema:
            type: string
        - in: query
          name: city
          schema:
            type: string
        - in: query
          name: subjectId
          schema:
            type: string
        - in: query
          name: includeDeleted
          schema:
            type: boolean
      responses:
        "200":
          description: Filtered classes